requirements.txt
# KynicOS — Requirements
# Compatible con Render (Python 3.11+)

# ── Core Framework ───────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# ── Telegram Bot ─────────────────────────────────────────────
python-telegram-bot>=21.0

# ── LLM Providers ───────────────────────────────────────────
groq>=0.11.0
anthropic>=0.34.0
openai>=1.30.0

# ── WhatsApp (Twilio) ────────────────────────────────────────
twilio>=9.1.0

# ── Payments (Stripe) — Fase 2 ───────────────────────────────
stripe>=10.0.0

# ── Database — Fase 2 ────────────────────────────────────────
asyncpg>=0.29.0
sqlalchemy>=2.0.0
redis>=5.0.0

# ── Vector DB / RAG — Fase 2 ─────────────────────────────────
# chromadb>=0.5.0        # self-hosted, activar en Fase 2
# sentence-transformers>=3.0.0  # activar para embeddings locales

# ── Config & Utils ───────────────────────────────────────────
pydantic>=2.4.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
httpx>=0.27.0
aiofiles>=23.2.0
tenacity>=8.2.0
loguru>=0.7.0
boto3>=1.29.0

# ── Dev / Test ───────────────────────────────────────────────
pytest>=7.4.0
pytest-asyncio>=0.21.0
